+++ gendarme/rules/Gendarme.Rules.Performance/AvoidUnneededFieldInitializationRule.cs
@@ -35,6 +35,7 @@ using Gendarme.Framework;
 using Gendarme.Framework.Engines;
 using Gendarme.Framework.Helpers;
 using Gendarme.Framework.Rocks;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Gendarme.Rules.Performance
 {
@@ -126,12 +127,32 @@ namespace Gendarme.Rules.Performance
           // we're more confident about the unneeded initialization
           // on static ctor, since another (previous) ctor, can't set
           // the values differently
-          Confidence c = method.IsStatic ? Confidence.High : Confidence.Normal;
+          Confidence c;
+          if (method.IsStatic)
+          {
+            c = Confidence.High;
+          }
+          else
+          {
+            if (IsGuiGeneratedCode(method, fr))
+              return RuleResult.DoesNotApply;
+            c = Confidence.Normal;
+          }
           Runner.Report(method, ins, Severity.Medium, c, frName);
         }
       }
 
       return Runner.CurrentRuleResult;
+    }
+
+#pragma warning disable IDE0079 // Remove unnecessary suppression
+    [SuppressMessage("Gendarme.Rules.Maintainability",
+                     "AvoidUnnecessarySpecializationRule",
+                     Justification = "That would be spurious generality")]
+    private static bool IsGuiGeneratedCode(MethodDefinition method, FieldReference fr)
+    {
+      return (string.Equals(fr.Name, "components", StringComparison.Ordinal)
+          && method.DeclaringType.IsDesignable());
     }
   }
 }